Remove flaky test
Depending on the order of the test suite execution we were having issues
as the compiled container for `CompilerTest` was using the same FQCN as
the `GeneratorTest`.

This makes sure we use different namespaces for the generated
containers, avoiding those conflicts.

It also plugs vfsStream to remove the file system dependency on
`CompilerTest`.

Running PHPUnit using the flag `--random-order-seed=1634593025` was
enough to reproduce the issue.